requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
async-timeout==5.0.1
    # via redis
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.4
    # via google-auth
google-api-core==2.29.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-storage==3.9.0
    # via spotify-history (pyproject.toml)
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.72.0
    # via google-api-core
idna==3.11
    # via requests
proto-plus==1.27.1
    # via google-api-core
protobuf==6.33.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
python-dotenv==1.2.1
    # via spotify-history (pyproject.toml)
redis==7.1.0
    # via spotipy
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-storage
    #   spotipy
rsa==4.9.1
    # via google-auth
spotipy==2.25.2
    # via spotify-history (pyproject.toml)
typing-extensions==4.15.0
    # via cryptography
urllib3==2.6.3
    # via
    #   requests
    #   spotipy
